+++ CHANGELOG.md
@@ -120,6 +120,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed database backups breaking with some languages
 - fixed when PlaceholderAPI variables contains dots
 - fixed quester name not support & as color code
+- fixed Region Objective listen to player teleport event
 ### Security
 - fixed issue, where objectives that count things are out of sync with the database. This has also affected BungeeCord support